Made index.js compatible with the Mattermost API V4
The V3 API of Mattermost will be deprecated in the next major Mattermost release, and can be manually disabled in Mattermost 4 already. As such, this needs to be upgraded. Besides updating it to the new API version, I changed the call from (the no longer existing) ``/users/initial_load``, to ``/system/ping``. As far as I could tell the original call wasn't used besides checking if it existed and contained a basic value, but contained a lot of not that useful information in the return call.  The ``/system/ping`` call should return the current status of the server better, and contains few information beyond that, making it slightly faster.