Fix notification bubble when being in multiple teams
It is possible to join multiple teams on the same Mattermost server.
Once one joins another team, Mattermost will show add a team sidebar.
When the user selected a different team, Mattermost shows the
notifications for other teams in this sidebar. To propagate the counter
to Franz we also need to extract the number of notifications for
different teams.